cambie para que imprima double

+++ ConvencionC/Ejercicio2/imprime.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
-extern void imprime(int a, float b, char * s);
+extern void imprime(int a, double b, char * s);
 
 int main() {
-	imprime(5,2.32f,"Holaaaa");
+	imprime(5,2.32,"Holaaaa");
 	return 0;
 }
